Updated Docker Compose with new structure & added gitignore.

+++ startup.sh
@@ -1,16 +1,32 @@
-apt-get update && \
+if [ ! -f /tmp/initializated ]; then
+    echo "Running one-time initialization command..."
+
+    touch /tmp/initializated
+
+    apt-get update && \
     apt-get install -y unzip && \
     apt-get clean
 
-unzip -qq /datadrive/tomcat/webapps/eisi.war -d /datadrive/tomcat/webapps/eisi/
+    unzip -qq /datadrive/tomcat/webapps/eisi.war -d /datadrive/tomcat/webapps/eisi/
 
-printf WAR Descomprimido
+    echo "WAR Descomprimido"
 
-rm -r /datadrive/tomcat/webapps/eisi/css/*
-rm -r /datadrive/tomcat/webapps/eisi/images/*
+    rm -rf /datadrive/tomcat/webapps/eisi/css/*
+    rm -rf /datadrive/tomcat/webapps/eisi/images/*
+    rm -rf /datadrive/tomcat/webapps/eisi/Archivos/*
+    rm -rf /datadrive/tomcat/webapps/eisi/ArchivosProfesores/*
+    mkdir /datadrive/tomcat/webapps/eisi/grupo
+    mkdir /datadrive/tomcat/webapps/eisi/WebProfesor
 
-cp -r /datadrive/tomcat/datos/css/* /datadrive/tomcat/webapps/eisi/css/
-cp -r /datadrive/tomcat/datos/images/* /datadrive/tomcat/webapps/eisi/images/
-cp -r /datadrive/tomcat/datos/archivos/* /datadrive/archivos/
+    cp -r /datadrive/tomcat/datos/css/* /datadrive/tomcat/webapps/eisi/css/
+    cp -r /datadrive/tomcat/datos/images/* /datadrive/tomcat/webapps/eisi/images/
+    cp -r /datadrive/tomcat/datos/Archivos/* /datadrive/tomcat/webapps/eisi/Archivos/
+    cp -r /datadrive/tomcat/datos/ArchivosProfesores/* /datadrive/webapps/tomcat/eisi/ArchivosProfesores/
+    cp -r /datadrive/tomcat/datos/WebProfesor/* /datadrive/tomcat/webapps/eisi/WebProfesor/
+    cp -r /datadrive/tomcat/datos/grupo/* /datadrive/tomcat/webapps/eisi/grupo/
+
+    cp -r /datadrive/tomcat/datos/archivosDatadrive/* /datadrive/archivos/
+    cp -r /datadrive/tomcat/datos/Profesores/* /datadrive/tomcat/webapps/
+fi
 
 catalina.sh run
